backend/ai-chat/requirements.txt
openai==1.54.0
psycopg2-binary==2.9.9
PyJWT==2.8.0